+++ src/components/ids-tree/ids-tree-shared.ts
@@ -16,7 +16,8 @@ const IdsTreeShared = {
     expandTarget: 'node',
     toggleCollapseIcon: 'plusminus-folder-closed',
     toggleExpandIcon: 'plusminus-folder-open',
-    toggleIconRotate: false
+    toggleIconRotate: false,
+    showExpandAndToggleIcons: false
   },
 
   SELECTABLE: ['single', 'multiple', 'none'],